+++ device/mpu9255/io/gyroscope/mpu9255_gyro.cpp
@@ -66,9 +66,9 @@ MPU9255_GyroRawData* MPU9255_Gyro::accumGyroOffset()
         usleep(1000);
     }
 
-    offset->x = tempGx / AVERAGE;
-    offset->y = tempGy / AVERAGE;
-    offset->z = tempGz / AVERAGE;
+    offset->x = int16_t(tempGx / AVERAGE);
+    offset->y = int16_t(tempGy / AVERAGE);
+    offset->z = int16_t(tempGz / AVERAGE);
 
     return offset;
 }
