+++ device/mpu9255/mpu9255_types.h
@@ -8,9 +8,9 @@ ACCELEROMETER DATA TYPES
 -----------------------------------------------------------------------*/
 struct MPU9255_AccelData
 {
-    float x = 0;
-    float y = 0;
-    float z = 0;
+    double x = 0;
+    double y = 0;
+    double z = 0;
 };
 
 /*-----------------------------------------------------------------------
@@ -18,9 +18,9 @@ MAGNETOMETER DATA TYPES
 -----------------------------------------------------------------------*/
 struct MPU9255_MagData
 {
-    float x = 0;
-    float y = 0;
-    float z = 0;
+    double x = 0;
+    double y = 0;
+    double z = 0;
 };
 
 /*-----------------------------------------------------------------------
@@ -28,9 +28,9 @@ GYROSCOPE DATA TYPES
 -----------------------------------------------------------------------*/
 struct MPU9255_GyroData
 {
-    float x = 0;
-    float y = 0;
-    float z = 0;
+    double x = 0;
+    double y = 0;
+    double z = 0;
 };
 
 #endif // MPU9255_TYPES_H
